+++ src/components/Duration.js
