+++ src/components/LegType/index.js
