+++ src/components/StyledText.js
